Fix backward scan flag for parallel queries (#158)
* Fix backward scan flag for parallel queries

Currently there is a bug in Babelfish that allows the backward scan flag
to be set for parallel queries. However, backward scans cannot be
parallelized. This bug results in assertion failures in the Postgres engine and
leads Babelfish queries to crash. In this commit, we fix the issue by
disallowing backward scan flag to be set when a query is in parallel
mode. This issue occurs during the implementation of full text search in
Babelfish. Fixing the issue unblocks full text search implementation.

Task: BABEL-4261, BABEL-4281
Signed-off-by: Robin Li <renyuli@amazon.com>