Block cross version and cross migration mode dump/restore (#204)
Description
Babelfish currently only supports dump and restore from databases with
same migration mode ( i.e. either singledb->singledb or multidb->multidb)
and same versions. This commit adds logic to dump a precheck so that the
restore process stops execution if the migration mode or version in the target
server is different than the one from which the dump file was created.

Task: BABEL-4055
Signed-off-by: Rishabh Tanwar ritanwar@amazon.com

Extensions PR: babelfish-for-postgresql/babelfish_extensions#1798